fix: build ready
